requirements.txt
asgiref==3.9.1
attrs==25.3.0
certifi==2025.7.9
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
django==5.2.4
django-allauth==65.10.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-redis==6.0.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
phonenumbers==9.0.9
psycopg2-binary==2.9.10
pyjwt==2.9.0
python-dotenv==1.1.1
pytz==2025.2
pyyaml==6.0.2
redis==6.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
sqlparse==0.5.3
typing-extensions==4.14.1
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
